feature/config-run-on-local-only (#291)
* chore: update roadmap

* chore: update roadmap

* chore: update roadmap

* feat: set config value temporary until restart on current node

* feat: set config value temporary until restart on current node

* feat: show database / current value in table

* feat: add tooltip on inactive value

* feat: docs

* feat: code review

* feat: code review

* fix: test

+++ src/main/java/sk/iway/iwcm/system/datatable/editorlocking/EditorLockingBean.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 public class EditorLockingBean {
 
     private int userId;
-    private int entityId;
+    private long entityId;
     private Date lastChange;
 
 }
